cleanup chain-storage
This is mainly a work related to removing old unused modules and merging
changes from different places.

- old chain-storage is now removed, because the traits it was providing
  are no longer used.
- chain-storage-sqlite-old is now called chain-storage instead.
- ChainStorageIndex was moved to the new chain-storage, but is not used
  for now. It needs to be properly integrated keeping the concurrency
  thing in mind.
