+++ Cargo.toml
@@ -7,7 +7,6 @@ members = [
     "chain-time",
     "chain-crypto",
     "chain-storage",
-    "chain-storage-sqlite-old",
     "chain-impl-mockchain",
     "cardano-legacy-address",
     "network-core",
