--- chain-storage-sqlite-old/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "chain-storage-sqlite-old"
-version = "0.1.0"
-authors = ["dev@iohk.io"]
-edition = "2018"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-chain-core = { path = "../chain-core" }
-rusqlite = { version = "0.22.0", features = ["bundled", "unlock_notify"] }
-thiserror = "1.0"
-
-criterion = { version = "0.3.0", optional = true }
-rand_core = { version = "0.5", features = ["getrandom"], optional = true }
-tempfile = { version = "3.1.0", optional = true }
-
-[features]
-with-bench = ["criterion", "tempfile", "rand_core"]
-
-[dev-dependencies]
-rand_core = { version = "0.5", features = ["getrandom"] }
-
-[[bench]]
-harness = false
-name = "storage"
-required-features = ["with-bench"]
