--- chain-storage-sqlite/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "chain-storage-sqlite"
-version = "0.1.0"
-authors = ["dev@iohk.io"]
-edition = "2018"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-chain-core = { path = "../chain-core" }
-chain-storage = { path = "../chain-storage" }
-r2d2 = { version = "0.8" }
-r2d2_sqlite = { version = "0.8" }
-thiserror = "1.0"
-
-[dependencies.rusqlite]
-version = "0.16.0"
-features = ["bundled"]
-
-[dev-dependencies]
-chain-storage = { path = "../chain-storage", features=["test-api"] }
-rand_core = "0.5"
